Rename files. Finish findUnfrozenVerts function.
