+++ meshChecker/src/pluginMain.cpp
@@ -1,9 +1,9 @@
-#include "topologyChecker.h"
+#include "meshChecker.h"
 #include <maya/MFnPlugin.h>
 
 MStatus initializePlugin(MObject mObj)
 {
-    MFnPlugin fnPlugin(mObj, "Michitaka Inoue", "0.0.1", "Any");
+    MFnPlugin fnPlugin(mObj, "Michitaka Inoue", "1.0", "Any");
     fnPlugin.registerCommand("checkMesh", MeshChecker::creator, MeshChecker::newSyntax);
     return MS::kSuccess;
 }
